backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
azure-cosmos==4.5.1
azure-storage-blob==12.19.0
python-dotenv==1.0.0
pyjwt==2.8.0
bcrypt==4.1.2
uuid
requests==2.31.0
opencensus-ext-azure==1.1.13
opencensus-ext-flask==0.8.0
opencensus-ext-logging==0.1.1